Stronghold - Routes: Refactor routes implementations, separating by concerns
